Provide a better way to run extra scripts from children images
The current way requires an `extra_scripts.sh` file to be present in on
children images.

With this approach a new folder `/docker-entrypoint.d` is created when
an image extends this one. The children image can then copy any shell or
python file that they want executed at the start of container in that
directory.

For instance consider the following setup:

```
cloud-mysite
├── .env
├── docker-entrypoint.d
│   └── 00_setup_validation.sh
│   └── 01_check_users.py
└── Dockerfile
```

The `Dockerfile` will look like:

```
FROM keitaro/ckan:2.8.1

[...]

COPY docker-entrypoint.d/* /docker-entrypoint.d/
```

The files will be executed in alphabetical order.