Fix check for `docker-entrypoint.d`
The entrypoint script checks for the existence of `docker-entrypoint.d` in the CWD, but then run the scripts in `/docker-entrypoint.d`. 

This causes the scripts to either not run at all or fail if an inheriting image changes the `WORKDIR`. 

This fix makes sure the directory is always expected in the filesystem root.